+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Fixed
 
 - Adopt new http proxy middleware
+- Remove `odata-version` header from request
 
 ## Version 1.12.8 - 2024-04-02